+++ app/controllers/questions_controller.rb
@@ -1,9 +1,10 @@
 class QuestionsController < ApplicationController
   before_action :authenticate_user!, except: [:index, :show]
-  expose :questions, ->{ Question.all }
+  expose :questions, -> { Question.all }
   expose :question
 
   def create
+    question.user = current_user
     if question.save
       redirect_to question_path(question), notice: 'Your question successfully created.'
     else
@@ -20,8 +21,9 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
+    return redirect_to question_path(question) unless current_user.questions.include?(question)
     question.destroy
-    redirect_to questions_path
+    redirect_to questions_path, notice: "Your question with title #{question.title} successfully delete."
   end
 
   private
